remove tree_to_string duplicity by moving tree_to_string to utils.py
